--- src/app/page/MakeMission/index.js
@@ -1,3 +1,0 @@
-import MakeMission from "./MakeMission";
-
-export default MakeMission;
