+++ src/app/page/MissionCreate/index.js
@@ -1,0 +1,3 @@
+import MissionCreate from "./MissionCreate";
+
+export default MissionCreate;
